Off-by-one error, plus return a size_t for a string length.

+++ sql/backends/monet5/UDF/pyapi/unicode.h
@@ -26,7 +26,7 @@
 	ascii: a pointer to a boolean, this is set to true if the string is ascii-encoded and to false otherwise
 */
 int utf8_strlen(const char *utf8_str, bool *ascii);
-int utf32_strlen(const Py_UNICODE *utf32_str);
+size_t utf32_strlen(const Py_UNICODE *utf32_str);
 
 //! Returns the length in bytes of a single utf8 character [1,2,3 or 4] based on the signature of the first byte, returns -1 if the character is not a valid utf8 character
 /* Arguments:
